requirements.txt
prefect
python-dotenv
google-cloud-storage